# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Streetly API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/marketplace-items:
    get:
      operationId: listMarketplaceItems
      tags: [marketplace]
      summary: List available marketplace items for a business (public storefront view)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marketplace items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceItem"
    post:
      operationId: createMarketplaceItem
      tags: [marketplace]
      summary: Owner/admin adds a marketplace item to a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceItemInput"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"
  /businesses/{businessId}/marketplace-items/manage:
    get:
      operationId: listMarketplaceItemsForManagement
      tags: [marketplace]
      summary: List all marketplace items for a business (owner/admin view, includes unavailable items)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marketplace items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceItem"
  /marketplace-items/{id}:
    patch:
      operationId: updateMarketplaceItem
      tags: [marketplace]
      summary: Owner/admin updates a marketplace item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceItemUpdateInput"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"
    delete:
      operationId: deleteMarketplaceItem
      tags: [marketplace]
      summary: Owner/admin deletes a marketplace item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item deleted
...
components:
  schemas:
...
    MarketplaceItem:
      type: object
      required: [id, businessId, name, price, isAvailable, createdAt]
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
    MarketplaceItemInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
    MarketplaceItemUpdateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        isAvailable:
          type: boolean
...